requirements.txt
streamlit
numpy
pandas
scikit-learn
tempfile
matplotlib.pyplot
tensorflow